refactor: decompose cli/doc.py into cohesive CLI submodules (Stage 5)
Pure extraction (no behaviour change): doc.py 1661 -> 1293 lines, with
helpers moved into new CLI-layer modules:
- cli/_doc_io.py: markdown/html input loaders
- cli/_doc_fetch.py: doc id resolution + block-fetch + not-found emitters
- cli/_doc_render.py: html/pdf/markdown render branches (extracted from
  get_cmd) + server-markdown emit + PDF image sanitizer
- cli/_doc_images.py: asset embed/download helpers

doc.py keeps every @app.command callback, the option enums, and dispatch.
Command surface verified identical to main (17 subcommands incl set/replace).
Left in the CLI layer by design: the helpers use GlobalOpts/typer, and the
pure block->markdown renderer already lives in mondo/docs.py.
